more fixes from review
