+++ src/containers/User/Detail/Detail.jsx
@@ -2,8 +2,8 @@ import { Button, classNames, Link } from "pi-ui";
 import React, { useEffect, useState } from "react";
 import { withRouter } from "react-router-dom";
 import ModalChangeUsername from "src/componentsv2/ModalChangeUsername";
-import { useQueryStringWithIndexValue } from "src/hooks/queryString";
 import useMediaQuery from "src/hooks/useMediaQuery";
+import useQueryStringWithIndexValue from "src/hooks/useQueryStringWithIndexValue";
 import styles from "./detail.module.css";
 import General from "./General.jsx";
 import { useChangeUsername, useUserDetail } from "./hooks";
