+++ src/containers/User/Detail/General.jsx
@@ -88,11 +88,11 @@ const General = ({
             <>
               {hasUserPaid(newuserpaywalltx, newuserpaywallamount) ? "Paid" : "Not paid"}
               {!hasUserPaid(newuserpaywalltx, newuserpaywallamount) && isAdmin ?
-                <Button loading={isApiRequestingMarkAsPaid} size="sm" onClick={openMarkAsPaidModal}>
+                <Button className="margin-top-s" loading={isApiRequestingMarkAsPaid} size="sm" onClick={openMarkAsPaidModal}>
                   Mark as paid
                 </Button>
                 : !hasUserPaid(newuserpaywalltx, newuserpaywallamount) &&
-                <Button size="sm">
+                <Button className="margin-top-s" size="sm">
                   Pay registration fee
                 </Button>
               }
